pyproject.toml
[project]
name = "politicoresto-admin-mcp"
version = "0.1.0"
description = "MCP admin server for the PoliticoResto Supabase backend (staging-only by default)."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Mickaël Labarrère" },
]
keywords = [
    "mcp",
    "model-context-protocol",
    "claude",
    "supabase",
    "politicoresto",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "pre-commit>=3.7",
    "bandit[toml]>=1.7",
    "build>=1.2",
]

[project.urls]
Homepage = "https://github.com/mlabarrere/politicoresto-admin-mcp"
Repository = "https://github.com/mlabarrere/politicoresto-admin-mcp"
Issues = "https://github.com/mlabarrere/politicoresto-admin-mcp/issues"
Changelog = "https://github.com/mlabarrere/politicoresto-admin-mcp/blob/main/CHANGELOG.md"

[project.scripts]
politicoresto-admin-mcp = "politicoresto_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/politicoresto_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "src/politicoresto_mcp",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "SKILL.md",
    ".env.example",
    "pyproject.toml",
]

# ---------------------------------------------------------------------------
# Ruff — lint + format (replaces black, isort, flake8)
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "C4",   # flake8-comprehensions
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific
    "BLE",  # flake8-blind-except
    "A",    # flake8-builtins
    "N",    # pep8-naming
    "S",    # flake8-bandit (light coverage, full bandit runs separately)
    "TID",  # flake8-tidy-imports
]
ignore = [
    "E501",   # Line too long — ruff format handles it
    "S101",   # Use of assert — fine in tests, overridden per-file
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert is expected in tests
    "S105",   # hardcoded password-like strings (fake keys)
    "S106",   # same
    "PLR2004",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Mypy — strict
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_unused_ignores = true
show_error_codes = true
files = ["src/politicoresto_mcp", "tests"]

[[tool.mypy.overrides]]
module = ["mcp.*", "mcp.server.*", "mcp.server.fastmcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
# Tests use fixtures and mocks that are awkward under strict mode.
disallow_untyped_decorators = false
warn_unused_ignores = false

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-m", "not integration",
]
markers = [
    "integration: hits a real Supabase project (opt-in only, skipped by default)",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:pydantic.*",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
branch = true
source = ["politicoresto_mcp"]
omit = ["src/politicoresto_mcp/__main__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_also = [
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# ---------------------------------------------------------------------------
# Bandit
# ---------------------------------------------------------------------------

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
# B101 (assert) is enforced by ruff in non-test code via S101.
skips = []
